Fix for pattern matching JSON where `request.json = true` is used
This fixes an incompatibility with newer versions of `request`, which has had about sixty releases since the pinned version, including some major security vulnerability patches. Security is not likely an issue in this repo as a devDependency, however when others using this framework update they will find certain pattern matchers not working.

This happened with `pkgcloud` in the PR at <https://github.com/pkgcloud/pkgcloud/pull/587>, where a work-around with heavily escaped strings can allow the patterns to be matched.

The specific case where patterns are not matching has been added as a unit test. It involves the documented behaviour of using `request` with `json: true` with `body: <JSON string>` rather than `json: <object value>`.